+++ NEWS.md
@@ -1,7 +1,7 @@
 # twidlr 0.0.0.9000
 
 * Added data.frame-first formula-second functions for:
-    * `stats::t.test, lm, glm, kmeans, prcomp`
+    * `stats::aov, t.test, lm, glm, kmeans, prcomp`
     * `xgboost::xgboost`
     * `glmnet::glmnet`
     * `rpart::rpart`